arreglos de marco legal, coordenadas tipo

+++ grails-app/domain/estacion/Coordenadas.groovy
@@ -5,8 +5,8 @@ class Coordenadas {
 
     static auditable = true
     Estacion estacion
-    Double coordenadasX
-    Double coordenadasY
+    BigInteger coordenadasX
+    BigInteger coordenadasY
 
     static mapping = {
         table 'cddn'
